# openapi.yml
openapi: 3.0.0
info:
  title: Priority Pilot API
  description: >-
    Vertrag für die Aufgaben-Priorisierung über einen gewichteten
    Abhängigkeitsgraphen. Aus dieser Datei werden via `openapi-typescript`
    sowohl die Server-Typen (`server/src/api.d.ts`) als auch die Client-Typen
    (`client/src/schema.d.ts`) erzeugt.
  version: 1.0.0
servers:
  - url: http://localhost:3000
paths:
...
  /categories:
    get:
      summary: Alle Kategorien auflisten
      description: >-
        Liefert die thematischen Kategorien des eingeloggten Nutzers. Kategorien sind
        die flache Ordnungsebene neben den Lebensbalance-Säulen: Sie gruppieren und
        filtern, ohne in Wert-, Score- oder Balance-Berechnung einzufließen.
      operationId: listCategories
      responses:
        '200':
          description: Liste aller Kategorien des Nutzers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      summary: Neue Kategorie anlegen
      description: >-
        Legt eine Kategorie für den eingeloggten Nutzer an. Der Name ist pro Nutzer
        eindeutig (409 bei Dublette), die Farbe stammt aus der festen Palette.
      operationId: createCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryCreate'
      responses:
        '201':
          description: Kategorie wurde angelegt
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          $ref: '#/components/responses/Conflict'
  /categories/{id}:
    parameters:
      - $ref: '#/components/parameters/CategoryId'
    patch:
      summary: Kategorie teilweise aktualisieren
      description: >-
        Ändert Name und/oder Farbe einer Kategorie. Bei Namenskonflikt 409, bei
        unbekannter oder fremder ID 404.
      operationId: updateCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryUpdate'
      responses:
        '200':
          description: Die aktualisierte Kategorie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    delete:
      summary: Kategorie löschen
      description: >-
        Löscht die Kategorie. Aufgaben und Serien bleiben erhalten und verlieren nur
        die Zuordnung (`categoryId` wird auf `null` gesetzt).
      operationId: deleteCategory
      responses:
        '204':
          description: Kategorie wurde gelöscht
        '404':
          $ref: '#/components/responses/NotFound'
...
  /tasks/suggest-pillars:
    post:
      summary: Säulen-Einzahlung per Mistral klassifizieren
      description: >-
        Schlägt anhand von Titel/Beschreibung/Kontext vor, auf welche Säulen eine
        Aufgabe einzahlt, samt Konfidenz (0–100) je Säule. Die Klassifikation
        erfolgt über die Mistral-API (`MISTRAL_API_KEY`, Modell via `MISTRAL_MODEL`,
        Default `mistral-small-latest`). Fehlt der Key oder gibt es keine Säulen,
        antwortet der Endpoint mit 503; bei Upstream-/Format-Fehlern mit 502.
      operationId: suggestPillars
      parameters:
        - name: provider
          in: query
          required: false
          description: >-
            Optionaler Provider-Pinning (#749/#951): Name eines konfigurierten
            Providers — überschreibt den aktiven Provider für diesen EINEN Aufruf.
            Fehlt der Parameter → der aktive Provider (`llm_providers`, #951).
            Ungültiger Wert → HTTP 400.
          schema:
            type: string
            maxLength: 64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SuggestPillarsInput'
      responses:
        '200':
          description: Vorgeschlagene Säulen-Einzahlungen mit Konfidenz
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuggestPillarsResult'
        '400':
          $ref: '#/components/responses/ValidationError'
        '502':
          $ref: '#/components/responses/BadGateway'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /tasks/parse-search:
    post:
      summary: Gesprochene/getippte Suchanfrage in Suchbegriff und Kategorie zerlegen
      description: >-
        Trennt eine frei formulierte Suchanfrage („offene Sachen zum Hausbau") in den
        reinen Suchbegriff und die gemeinte Kategorie des Nutzers. Die Kategorien des
        eingeloggten Nutzers gibt der Server dem Modell vor; erkennt es keine, bleibt
        `categoryId` leer. Fehlt der LLM-Key, antwortet der Endpoint mit 503; bei
        Upstream-/Format-Fehlern mit 502.
      operationId: parseSearchQuery
      parameters:
        - name: provider
          in: query
          required: false
          description: >-
            Optionaler Provider-Pinning (#749/#951): Name eines konfigurierten
            Providers — überschreibt den aktiven Provider für diesen EINEN Aufruf.
          schema:
            type: string
            maxLength: 64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ParseTaskInput'
      responses:
        '200':
          description: Aus der Anfrage extrahierter Suchbegriff und Kategoriefilter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParsedSearch'
        '400':
          $ref: '#/components/responses/ValidationError'
        '502':
          $ref: '#/components/responses/BadGateway'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /tasks/parse-text:
    post:
      summary: Task-Freitext per Mistral in strukturierte Felder parsen
      description: >-
        Extrahiert aus einem frei formulierten Text (Schnellerfassung, #235) die
        strukturierten Felder einer Aufgabe (Titel plus optional Beschreibung,
        Priorität, geschätzter Aufwand und Deadline). Das Parsing erfolgt über die
        Mistral-API (`MISTRAL_API_KEY`, Modell via `MISTRAL_MODEL`, Default
        `mistral-small-latest`). Fehlt der Key, antwortet der Endpoint mit 503; bei
        Upstream-/Format-Fehlern mit 502.
      operationId: parseTaskText
      parameters:
        - name: provider
          in: query
          required: false
          description: >-
            Optionaler Provider-Pinning (#749/#951): Name eines konfigurierten
            Providers — überschreibt den aktiven Provider für diesen EINEN Aufruf.
            Fehlt der Parameter → der aktive Provider (`llm_providers`, #951).
            Ungültiger Wert → HTTP 400.
          schema:
            type: string
            maxLength: 64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ParseTaskInput'
      responses:
        '200':
          description: Aus dem Text extrahierte, strukturierte Task-Felder
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParsedTask'
        '400':
          $ref: '#/components/responses/ValidationError'
        '502':
          $ref: '#/components/responses/BadGateway'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  parameters:
...
    CategoryId:
      name: id
      in: path
      required: true
      description: ID der Kategorie
      schema:
        type: integer
  responses:
    ValidationError:
      description: Ungültige Eingabe
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Ressource nicht gefunden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: Konflikt — die Abhängigkeit würde einen Zyklus erzeugen
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalServerError:
      description: Interner Serverfehler (z. B. Datenbankfehler)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    BadGateway:
      description: Fehler beim Upstream-Dienst (Mistral nicht erreichbar/ungültige Antwort)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ServiceUnavailable:
      description: Dienst nicht konfiguriert (z. B. fehlender API-Key oder keine Säulen)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Category:
      type: object
      description: >-
        Thematische Kategorie einer Aufgabe/Serie — die flache Ordnungsebene neben den
        Lebensbalance-Säulen. Höchstens eine je Aufgabe, ohne Anteil und ohne Wirkung
        auf Wert, Score oder Balance; sie dient dem Gruppieren und Filtern.
        Nutzer-eigene Stammdaten, Name pro Nutzer eindeutig.
      required:
        - id
        - name
        - color
      properties:
        id:
          type: integer
        name:
          type: string
          maxLength: 40
          description: 'Name der Kategorie (nicht leer, z. B. Hausbau, Steuer, Verein)'
        color:
          $ref: '#/components/schemas/CategoryColor'
    CategoryColor:
      type: string
      description: >-
        Badge-Farbe aus der festen Palette. Bewusst keine freie Farbwahl: Die Werte
        sind auf die Design-Sprache abgestimmt, die Textfarbe rechnet `KolBadge` selbst
        aus. Spiegel von `server/src/models/categoryColors.ts`.
      enum:
        - '#b42318'
        - '#b54708'
        - '#8a6100'
        - '#1a7f37'
        - '#0e7490'
        - '#1064d0'
        - '#6941c6'
        - '#475467'
    CategoryCreate:
      type: object
      description: Eingabe zum Anlegen einer neuen Kategorie.
      required:
        - name
        - color
      properties:
        name:
          type: string
          maxLength: 40
          description: Name der Kategorie (nicht leer)
        color:
          $ref: '#/components/schemas/CategoryColor'
    CategoryUpdate:
      type: object
      description: Eingabe zum teilweisen Aktualisieren einer Kategorie. Mindestens ein Feld muss gesetzt sein.
      properties:
        name:
          type: string
          maxLength: 40
          description: Neuer Name der Kategorie (nicht leer, falls gesetzt)
        color:
          $ref: '#/components/schemas/CategoryColor'
...
    SuggestPillarsInput:
      type: object
      description: Eingabe für die Säulen-Klassifikation einer Aufgabe.
      required:
        - title
      properties:
        title:
          type: string
          description: Titel der Aufgabe (Pflichtfeld, nicht leer)
        description:
          type: string
          description: Optionale ausführliche Beschreibung der Aufgabe
        context:
          type: string
          description: Optionaler Kontext, z. B. zu abhängigen Aufgaben
    PillarSuggestion:
      type: object
      description: Eine vorgeschlagene Säulen-Einzahlung mit Konfidenz.
      required:
        - pillarId
        - confidence
      properties:
        pillarId:
          type: integer
          minimum: 1
          description: ID der Säule, auf die die Aufgabe einzahlt
        confidence:
          type: number
          format: float
          minimum: 0
          maximum: 100
          description: Konfidenz in % (0–100), dass die Aufgabe auf diese Säule einzahlt
    SuggestPillarsResult:
      type: object
      description: Ergebnis der Säulen-Klassifikation.
      required:
        - suggestions
      properties:
        suggestions:
          type: array
          description: Vorgeschlagene Säulen-Einzahlungen, nach pillarId sortiert
          items:
            $ref: '#/components/schemas/PillarSuggestion'
    ParseTaskInput:
      type: object
      description: Eingabe für die Task-Schnellerfassung — frei formulierter Text (#235).
      required:
        - text
      properties:
        text:
          type: string
          maxLength: 2000
          description: Frei formulierter Text, aus dem die Task-Felder extrahiert werden (nicht leer)
    ParsedTask:
      type: object
      description: >-
        Aus einem Freitext extrahierte, strukturierte Task-Felder (#235). Nur `title`
        ist garantiert gesetzt; die restlichen Felder liefert das Modell nur, wenn der
        Text sie hergibt.
      required:
        - title
      properties:
        title:
          type: string
          description: Kurzer, prägnanter Titel der Aufgabe (Pflichtfeld, nicht leer)
        description:
          type: string
          description: Optionale ausführliche Beschreibung der Aufgabe
        priority:
          type: integer
          minimum: 1
          maximum: 5
          description: Optionale Priorität 1–5 (1 = niedrig, 5 = hoch)
        estimatedEffort:
          type: number
          format: float
          minimum: 0
          description: Optionaler geschätzter Aufwand in Personentagen (z. B. 0.25 ≈ 2 h)
        deadline:
          type: string
          format: date-time
          description: Optionale Deadline als ISO-8601-Datum/Zeit-String
        isSeries:
          type: boolean
          description: >-
            Optionale Serien-Erkennung (#1310): `true`, wenn der Text einen wiederkehrenden
            Termin beschreibt — das Formular startet dann im Serien-Modus.
        address:
          type: string
          description: >-
            Optionale Ortsangabe aus dem Freitext (#1310) als Adresstext; reiner Freitext,
            ohne Koordinaten (kein Geocoding im Parsing).
        checklist:
          type: array
          items:
            type: string
            minLength: 1
            maxLength: 255
          description: >-
            Optionale, im Freitext aufgezählte Einzelpunkte (#1310) — je Eintrag ein
            Checklisten-Eintrag im Formular.
        categoryId:
          type: integer
          description: >-
            Optionale Kategorie: ID einer Kategorie des Nutzers, wenn der Text thematisch
            eindeutig zu einer passt. Fehlt, wenn keine passt oder keine angelegt sind.
    ParsedSearch:
      type: object
      description: >-
        Aus einer frei formulierten Suchanfrage extrahierter Filter. Beide Felder sind
        optional: `text` fehlt, wenn die Anfrage nur eine Kategorie nennt, `categoryId`
        fehlt, wenn keine Kategorie gemeint ist.
      properties:
        text:
          type: string
          description: Reiner Suchbegriff ohne den Kategorie-Anteil der Anfrage
        categoryId:
          type: integer
          description: ID der gemeinten Kategorie des Nutzers
...
    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: Fehlerbeschreibung
...